fix(risks): hide mutation controls for read-only viewers in risk detail
- Extract canEditRisk/canDeleteRisk into a shared useRiskPermissions
  hook (moved out of riskRowActions.tsx) so the same access decision
  drives both the header's Edit/Delete menu and the detail content.
- Gate document upload, 'Add update', and each update entry's
  Edit/Delete menu behind canEditRisk so read-only viewers no longer
  see mutation controls that would fail at the API.
- Move the risk owner/owner labels to render after the name
  ('Name - Risk owner') instead of before, in the detail header pill.
